Supports covering messages signaling errors
**Issue description**

Before this commit, the keywords of messages signaling errors could not be covered.

e.g. when analyzing the coverage of the following collaboration

```
"Original Code"

self signal: 'oops!'
```

the only parse nodes that were covered were `self` and `'oops!'`.
The `signal:` parse node could not be covered as the error raised prevented the execution of the code coverage tracking code:

```
"Original code with code coverage tracing"

((self) coverAll: {...})
    signal: ('oops!' coverAll: {...}}) coverAll: {...}
```

Note the outer `coverAll:` message send was never executed because the error raised from `signal:` prevented it.

**Solution**

This commit changes the way message sends are analyzed. Instead of sending `coverAll:` message to the message result,
the whole message is wrapped in a block. Code coverage is analyzed by sending the `valueCoveringAll:` to that block.

```
"Original code with the new code coverage tracing from this commit"

[
    (self) coverAll: {...})
        signal: ('oops!' coverAll: {...}}
] valueCoveringAll: {...}
```

The `valueCoveringAll:` message will take care of collecting the code coverage information for its statements regardless
if they succeed or not.