Covers boolean instance variables
A boolean instance variable is considered 100% covered when the variable has been assigned to true and false al least once in the compiled method.
When an instance variable is assigned to both boolean and non-boolean objects, it will not be considered a boolean instance variable.

Note: this commit does not add support for covering boolean instance variables across different compiled methods.