Covers boolean class variables
A boolean class variable is considered 100% covered when the variable has been assigned to true and false al least once in the compiled method.
When an v variable is assigned to both boolean and non-boolean objects, it will not be considered a boolean class variable.

Note: this commit does not add support for covering boolean class variables across different compiled methods.