Covers collaborators in cascade message sends
Thanks @hernanwilkinson for discovering the bug and providing a test for it!

This commit adds support for covering collaborators in cascade message sends.
The previous implementation used the whole message send range to cover it
(self consumeNextSourceRangeFor: aMessageNode) which included both the message keywords
and collaborators.
Additionally, the collaborators themselves were not traced, so the combination of this
two factors gave the false illusion that they were traced.

To fix the bug, cascade message sends collaborators are traced in the same way regular
message sends collaborators are.
With that in place, the only source ranges remaining to be covered were the ones for the
cascade message keywords. This was done in the same way regular message send keywords are traced.

Finally, the #cascadeNodeWithMessageSendReceiver example was refactored to not include cascade message
sends with collaborators (as this scenario was covered in a new test introduced in this commit).