Don't filter non-covered classes in the Browser
This bug caused pure abstract classes not to be displayed in the Code
Coverage Browser.

By definition, abstract methods are filtered out when installing compiled
method analyzers because they should never be executed and their coverage
should always be 0%. If we don't do this, they will lower the average
coverage ratio of the class/package they belong to.

Because of that, pure abstract classes (defining only abstract methods)
didn’t show in the Code Coverage Browser because they were not analyzed.

If a subclass of an abstract class happened to be analyzed, it was
displayed indented in the Code Coverage Browse, but the parent
class was missing. This made the class look like they subclassed the
class that happened to be rendered before them in the class list.

Thanks for reporting this bug @hernanwilkinson!