Tracks method execution in the traced source code
This commit moves the responsibility to track compiled methods execution
from `CompiledMethodCoverageAnalyzer` to `CodeCoverageSourceCodeGenerator`.

Instead of keeping track of whether a compiled method was executed in
`CompiledMethodCoverageAnalyzer` (acts as a method wrapper), an extra
parse node was added to the start of traced method to register in the
coverage report builder when a method is executed.

This cleans up the responsibilities and moves all coverage tracking to
`CodeCoverageSourceCodeGenerator`.