Reports code coverage to many report builders
This commit allows compiled method coverage analyzers to report code
coverage information to more than one report builder.
To do this the CompiledMethodCoverageTracker hierarchy was reified.

Instead of binding an instance of `CompiledMethodCoverageReportBuilder`
in the traced compiled methods, an instance of the new `CompiledMethodCoverageNotifier`
class is bound instead.
This class is a composite of `CompiledMethodCoverageTracker` instances,
allowing the same message to be dispatched to its children.
In this way, we can support notifying code coverage information from
traced compiled methods to many report builders.

This is needed to add support for analyzing a compiled method with more
than one `CodeCoverageAnalyzer`. In particular, this is interesting to allow
Code Coverage to analyze itself.