Supports multiple CompiledMethodCoverageNotifiers
This commit generates unique coverage tracker literal names in
`CodeCoverageSourceCodeGenerator` to support multiple
`CodeCoverageAnalyzer`s analyzing the same compiled method.

If this was not done, then the `CompiledMethodCoverageNotifier` instance
setup by the first `CompiledMethodCoverageAnalyzer` instance on a
compiled method was overwritten when another
`CompiledMethodCoverageAnalyzer` was installed, causing 0% coverage to
be reported on the first one (because it never received any coverage
notifications).