Displays coverage percentages on system categories
This commit fixes a bug that prevented code coverage percentages from
being displayed in system categories.

The bug was caused by changes in the Cuis base image: the component that
displays system categories was changed from `PluggableListMorph` to
`HierarchicalListMorph`.

The fix involved hacking `CodeCoverageResultItem` to make it work for
system categories, while avoiding to change the base image code as much
as possible.