Supports decision coverage on blocks with returns
This commit supports decision coverage on blocks that return from the
method they created from.

Before this commit, the result of a decision selector was not tracked if
the block evaluated by it returned from the method it was created from.
This means that even if the decision selector was evaluated on both
branches, the branch with the return was not tracked and the selector
was covered 50% instead of 100%.