Fixes browsing the message category list
This commit fixes the message category list in the code coverage browser.
Clicking on some categories did not work and the message list was empty.
The bug was caused when the message category list contents were changed
to be instances of `CodeCoverageResultItem`.
This broke the equality tests in #messageList and returned an empty
collection.
The bug was fixed by unwrapping the item in `CodeCoverageResultItem` in
`#messageList` (copied from the superclass).