Evaluates message receivers only once
This commit fixes a bug that caused message receivers to be evaluated
twice.
The bug was caused when message node were rewritten to call
`#coverAll:evaluating:thatSends:to:`: the message receiver node was used
for creating the block to be passed to the `evaluating:` and `to:`
keywords.

If the receiver happened to be another message send (say the message was
`(Set new) add: 1`), then `Set new` would be evaluated twice.

To fix the bug, message nodes are rewritten and all reference to their
receivers are done using a temporary variable. This way receivers are
evaluated only once.